TaskClosePanel: fix Close button bug
Eliminate the "Click()" function for the ClosePanel
that was testing for task changes.
Moved the test for task modification in the Close Panel's
Show() method where it works correctly because
the previous panel has been hidden and updated
the temporary task with any changes.

Fixes bug 1619 where changes to the Properties panel
were not saved if the Close button was clicked from
the Properties panel.