InfoBoxWindow.cpp: fix invalid type conversion
It was converting int x and int y to unsigned before
checking that the values were positive.
Their values could be negative because we capture the mouse
and the Mouse Up coordinates could be negative.
It was working because a small negative number converts
to a huge positive number, so the condition was evaluated
to true.  But int should not be converted to unsigned if it has
a negative value.